A little change in Region interface and how we handle the stored views.
We do not use Region to access particular components of the views. Instead,
we get the whole big fat view from Region, and access it from the Kokkos
parallel loops.

I did this because I'm not sure how correct would be to access an object
of type Region on a Cuda memory space. I suspect it would be not correct,
but I may be wrong.

One thing I need to understand: is it expensive to create a subview as a
separate view? What if I use that several times? When is the crossover?