Fixed a logic bug in preq_omega_ps
We cannot use sphere routines inside TeamThreadRange, since they are 'collective'
on the thread. Moreover, in preq_omega_ps, to save memory, we are recycling
'pressure', which is also the input to gradient_sphere. As a result, before
the fix, gradient_sphere was not using (for each level) ALL the original values
of 'pressure', but a mix of original and updated values.